+++ src/features/types.ts
@@ -1,3 +1,5 @@
+import type {Socket} from "node:net";
+
 export type ColorDepth = 1 | 3 | 4 | 8 | 24;
 
 export interface Features {
@@ -8,3 +10,10 @@ export interface Features {
 	scroll     : boolean;
 	terminal   : boolean;
 }
+
+export type Args =
+	| [enabled    : boolean,           stream?: NodeJS.WriteStream]
+	| [colorDepth : ColorDepth,        stream?: NodeJS.WriteStream]
+	| [features   : Partial<Features>, stream?: NodeJS.WriteStream]
+	| [socket     : Socket]
+	| [stream?    : NodeJS.WriteStream];