+++ src/sgr/style.ts
@@ -1,21 +1,38 @@
-import {makeSGR} from "./sgr";
+import {TypedObject} from "../typed-object";
+import {lazy}        from "./lazy";
 
-export function makeStyle(enabled = true) {
-	const sgr = makeSGR(enabled);
+import type {ChainKey, FormatBuilder, Style} from "./types";
 
-	return {
-		bold            : sgr(1, 22, true),
-		dim             : sgr(2, 22, true),
-		italic          : sgr(3, 23),
-		underline       : sgr(4, 24),
-		inverse         : sgr(7, 27),
-		hidden          : sgr(8, 28),
-		strikethrough   : sgr(9, 29),
-		doubleUnderline : sgr(21, 24),
-		frame           : sgr(51, 54),
-		encircle        : sgr(52, 54),
-		overline        : sgr(53, 55),
-	};
+const propParams = TypedObject.entries<Record<keyof Style, SkipFirst<Parameters<FormatBuilder>>>>({
+	bold            : [ 1, 22, true],
+	dim             : [ 2, 22, true],
+	italic          : [ 3, 23],
+	underline       : [ 4, 24],
+	inverse         : [ 7, 27],
+	hidden          : [ 8, 28],
+	strikethrough   : [ 9, 29],
+	doubleUnderline : [21, 24],
+	frame           : [51, 54],
+	encircle        : [52, 54],
+	overline        : [53, 55],
+});
+
+export function makeStyle(
+	keys          : ReadonlySet<ChainKey>,
+	makeFormatter : FormatBuilder,
+	enabled       : boolean,
+): Style<any> {
+	const rtn = {} as Style<any>;
+
+	for (const [key, args] of propParams) {
+		if (keys.has(key)) {
+			lazy.add(rtn, key, () => {
+				const newKeys = new Set(keys);
+				newKeys.delete(key);
+				return enabled ? makeFormatter(newKeys, ...args) : makeFormatter(newKeys, 0, 0);
+			});
+		}
+	}
+
+	return rtn;
 }
-
-export type Style = ReturnType<typeof makeStyle>;